mobile and desktop into separate components
