+++ src/components/NewLocationPage/SummaryStats/Stat.tsx
@@ -1,22 +1,11 @@
 import React from 'react';
-import MetricMeasure from './MetricMeasure';
-import StatValue from './StatValue';
-import {
-  MetricLabel,
-  StatContent,
-  Row,
-  StyledChevron,
-  MobileOnly,
-  DesktopOnly,
-} from './SummaryStats.style';
+import MobileStat from './MobileStat';
+import DesktopStat from './DesktopStat';
+import { MobileOnly, DesktopOnly } from './SummaryStats.style';
 import { formatValue, getLevelInfo } from 'common/metric';
 import { Metric } from 'common/metricEnum';
 import { useBreakpoint } from 'common/hooks';
 import { getMetricNameForStat, metricMeasureText } from './utils';
-
-const MetricName: React.FC<{ metricName: string }> = ({ metricName }) => {
-  return <MetricLabel>{metricName}</MetricLabel>;
-};
 
 const Stat: React.FC<{ metric: Metric; value: number }> = ({
   metric,
@@ -28,37 +17,24 @@ const Stat: React.FC<{ metric: Metric; value: number }> = ({
   const metricName = getMetricNameForStat(metric);
   const isMobile = useBreakpoint(600);
 
+  const statProps = {
+    levelInfo,
+    formattedValue,
+    showMetricMeasureText,
+    metricName,
+    isMobile,
+    metric,
+  };
+
   return (
-    <StatContent>
-      <Row>
-        <MetricName metricName={metricName} />
-        <MobileOnly>
-          {showMetricMeasureText && (
-            <MetricMeasure
-              text={metricMeasureText[metric]}
-              isMobile={isMobile}
-            />
-          )}
-        </MobileOnly>
-        <DesktopOnly>
-          <StyledChevron />
-        </DesktopOnly>
-      </Row>
-      <Row>
-        <StatValue value={formattedValue} iconColor={levelInfo.color} />
-        <MobileOnly>
-          <StyledChevron />
-        </MobileOnly>
-        <DesktopOnly>
-          {showMetricMeasureText && (
-            <MetricMeasure
-              text={metricMeasureText[metric]}
-              isMobile={isMobile}
-            />
-          )}
-        </DesktopOnly>
-      </Row>
-    </StatContent>
+    <>
+      <MobileOnly>
+        <MobileStat {...statProps} />
+      </MobileOnly>
+      <DesktopOnly>
+        <DesktopStat {...statProps} />
+      </DesktopOnly>
+    </>
   );
 };
